added a test, created a util file
