+++ godemon.go
@@ -9,70 +9,6 @@ import (
 	"time"
 )
 
-type (
-	Qnode[T any] struct {
-		Value T
-		Next  *Qnode[T]
-		Prev  *Qnode[T]
-	}
-	Queue[T any] struct {
-		Head   *Qnode[T]
-		Tail   *Qnode[T]
-		Length int8
-	}
-	Qinterface[T any] interface {
-		Enqueue(v T)
-	}
-	FileTreeNode struct {
-		Value    os.FileInfo
-		Path     string
-		Children []*FileTreeNode
-	}
-	FileChecking[T any] interface {
-		BFS(v os.FileInfo)
-		Error() error
-	}
-)
-
-func (n *FileTreeNode) Error() error {
-	return fmt.Errorf("There is an error in godemon 𓁹‿𓁹")
-}
-
-func NewQueue[T any]() *Queue[T] {
-	return &Queue[T]{}
-}
-
-func (q *Queue[T]) Enqueue(v T) {
-	newNode := Qnode[T]{Value: v}
-	if q.Head == nil {
-		q.Head = &newNode
-		q.Tail = &newNode
-	} else {
-		q.Tail = &newNode
-		q.Tail.Next = &newNode
-	}
-	q.Length++
-}
-
-func (q *Queue[T]) Dequeue() *T {
-	if q.Head == nil {
-		return nil
-	}
-
-	head := q.Head
-	q.Head = q.Head.Next
-
-	head.Next = nil
-	q.Length--
-	return &head.Value
-}
-
-func NewFileNode(relPath string) *FileTreeNode {
-	value, err := os.Stat(relPath)
-	check(err)
-	return &FileTreeNode{Value: value, Path: relPath}
-}
-
 func (n *FileTreeNode) BFS(t time.Time) *FileTreeNode {
 
 	q := NewQueue[FileTreeNode]()
@@ -102,13 +38,6 @@ func (n *FileTreeNode) BFS(t time.Time) *FileTreeNode {
 	}
 
 	return n
-}
-
-func (v *FileTreeNode) changed(t time.Time) bool {
-	if v.Value.ModTime() != t {
-		return true
-	}
-	return false
 }
 
 func IgnoreDirs(ignoreDirs map[string]bool) error {
