+++ godemon_test.go
@@ -1,0 +1,3 @@
+package main
+
+//To be written...
