+++ util.go
@@ -1,6 +1,76 @@
 package main
 
-import "log"
+import (
+	"fmt"
+	"log"
+	"os"
+	"time"
+)
+
+type (
+	Qnode[T any] struct {
+		Value T
+		Next  *Qnode[T]
+		Prev  *Qnode[T]
+	}
+	Queue[T any] struct {
+		Head   *Qnode[T]
+		Tail   *Qnode[T]
+		Length int8
+	}
+	Qinterface[T any] interface {
+		Enqueue(v T)
+		Dequeue() *T
+	}
+	FileTreeNode struct {
+		Value    os.FileInfo
+		Path     string
+		Children []*FileTreeNode
+	}
+	FileChecking[T any] interface {
+		BFS(t time.Time) *FileTreeNode
+		Error() error
+	}
+)
+
+func (n *FileTreeNode) Error() error {
+	return fmt.Errorf("There is an error in godemon 𓁹‿𓁹")
+}
+
+func NewQueue[T any]() *Queue[T] {
+	return &Queue[T]{}
+}
+
+func (q *Queue[T]) Enqueue(v T) {
+	newNode := Qnode[T]{Value: v}
+	if q.Head == nil {
+		q.Head = &newNode
+		q.Tail = &newNode
+	} else {
+		q.Tail = &newNode
+		q.Tail.Next = &newNode
+	}
+	q.Length++
+}
+
+func (q *Queue[T]) Dequeue() *T {
+	if q.Head == nil {
+		return nil
+	}
+
+	head := q.Head
+	q.Head = q.Head.Next
+
+	head.Next = nil
+	q.Length--
+	return &head.Value
+}
+
+func NewFileNode(relPath string) *FileTreeNode {
+	value, err := os.Stat(relPath)
+	check(err)
+	return &FileTreeNode{Value: value, Path: relPath}
+}
 
 // check() takes in error and checks if it's nil or not.
 // It takes the boilerplate and basically transforms into a very easy function.
@@ -9,4 +79,13 @@ func check(e error) {
 	if e != nil {
 		log.Fatalf("Some error occured: %v\n", e)
 	}
+}
+
+// changed(t time.Time) (bool) tracks actual changing time of the file.
+// "t" represents initial time that everything tracks with.
+func (v *FileTreeNode) changed(t time.Time) bool {
+	if v.Value.ModTime() != t {
+		return true
+	}
+	return false
 }
